ARM: omap4+: mux: Forbid omap_mux_init_signal to mux with WAKEUP_EN
omap_mux_init_signal() used with OMAP_WAKEUP_EN will set the corresponding
WAKEUPENABLE bit for the PAD.

But we miss the WAKEUPENABLE/WAKEUPEVENT bits management done when the
omap_mux_init_gpio() or omap_hwmod_mux_init() function is used.

The consequence is that the WAKEUPEVENT bit will never be cleared for this
PAD and so the Daisy Chain IO PAD wakeup mechanism won't generate a wakeup
anymore.

To avoid it, this patch forbids any configuration including WAKEUP_EN.

Change-Id: I916b8fa4ebcbb7dc0fbeca4d8cb881df6de297e0
Signed-off-by: Cedric Poignet <c-poignet@ti.com>