Input: omap4-keypad: fix errata: i689:Keyboard Key Up Event Can Be Missed
Inherited from commit d4d07c95969da746524e6ba34df786f9ac746db2 in p-android-omap-3.0
Reworked to adapt to p-linux-omap-3.4 as part for migration from k3.0 to k3.4

Silicon Errata #i689

Long key press is not needed on android since it
is simulated on user space. Extra interrupts may
leave userspace confused.

Use bitfiled instead of hardcoded values to set
KBD_CTRL, and change PVR divisor to 0x6, which
would allow a debouncing time of ~10ms.

Set mpu irq to level instead of edge, since if mpu
is in low power a edge detection may be lost if
the event is a wkup event.

Signed-off-by: Andrii Danylov <x0158318@ti.com>
Signed-off-by: Axel Haslam <axelhaslam@ti.com>

Errata Description:

When a key is released for a time shorter than the debounce time,
in-between 2 key press (KP1 and KP2), the keyboard state machine will go to
idle mode and will never detect the key release (after KP1,and also after KP2),
and thus will never generate a new IRQ indicating the key release.
From the operating system standpoint, only a key press as been detected,
and the key will keep on being printed on the screen until another or
the same key is pressed again.When the failing state has been reached,
the KBD_STATEMACHINE register will show "idle state" and the KBD_FULLCODE register
won't be empty, this is the signature of the bug.There is no impact on the power
consumption of the system as the state machine goes to IDLE state.

WORKAROUND
First thing is to program the debounce time correctly:
If X (us) is the maximum time of bounces when a key is pressed or released,
and Y (us) is the minimum time of a key release that is expected to be detected,
then the debounce time should be set to a value inbetween X and Y.

In case it is still possible to get shorter than debounce time key-release events,
then the only solution is to implement a software workaround:

Before printing a second character on the screen, the software must check if
the keyboard has hit the failing condition (cf signature of the bug above)
or if the key is still really pressed and then take the appropriate actions.

Silicon revisions impacted: OMAP4430 ES 1.0,2.0,2.1,2.2,2.3
                            OMAP4460 ES 1.0,1.1
                            OMAP5430 ES 1.0

Change-Id: I59547ed9751e487eb95d6bd2bf2dbe71401197ec
Signed-off-by: Ranganath Krishnan <ranganath@ti.com>