ARM: OMAP5: PM: export omap_bus_sync when erratum is disabled
Export the symbol omap_bus_sync as it is exported when
CONFIG_OMAP4_ERRATA_I688 is defined. However any module
that uses this symbol will fail to function with it disabled.
So, ensure it is continued to be exposed with CONFIG_OMAP4_ERRATA_I688
disabled as well, even though the implementation is in sleep
assembly code.

Change-Id: Iaf00afc35ed80e26a8fcae5b058eded44fc5727c
Signed-off-by: Nishanth Menon <nm@ti.com>