ARM: OMAP: hwmod: Fix omap_hwmod_disable_clocks API
When omap_hwmod_disable_clocks() API is calling it disables the clock
first and then disables module, but this is wrong, because
the module must be in idle mode before disabling any parents
clocks. Otherwise, the parent clock might be disabled before
the module transition is done, and thus will prevent the transition to
complete properly.

Hence disable the module before cut its clock.

Change-Id: Ie25359e8327e716e4d19d49f36456cea1eb6f370
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>