OMAP5: GPIO: WA for glitch on Modem reset GPIO pin
In OMAP5, GPIO pin 70 is the reset pin for the modem IC.
Due to h/w bug OMAP5430-1.0BUG01667 in ES 1.0 sometimes there is a
glitch observed while going to OFF mode resulting in reset of these
modules. This patch is a WA for this issue where the pin directions
are changed to input.
Along with this the MUX also needs be modified.
This needs to be reverted since the issue is not observed on ES 2.0

Change-Id: I0b0a72c0e0ab079ce094f6a438df4e96e6e9e4b9
Signed-off-by: Cedric Poignet <c-poignet@ti.com>