OMAP5 sEVM: Modem: Fix Modem reset mux setting
Due to GPIO bug OMAP5430-1.0BUG01667 in OMAP5 ES1.0, a glitch is
sometimes observed on the Modem reset GPIO pins.
WA is to change direction to input. For this WA to be effective
the mux configuration needs to enable INPUT_PULLUP. This will
have no effect when configured in output mode but is required when
pin is changed to input during OSWR/OFF mode.
This along with "OMAP5: GPIO: WA for glitch on Modem reset GPIO pin"
can be reverted when this is resolved in ES 2.0.

Change-Id: Ic45d44235bd896534e903648f4e67200b4f3f5fd
Signed-off-by: Cedric Poignet <c-poignet@ti.com>