ARM: OMAP5: CPUIdle : Call Bandgap-Thermal notifiers during idle path
To let PD_CORE transitions to low power states, Bandgap-Thermal
driver should call its callback notifiers in order to idle the
CORE during cpuidle.

Bandgap-Thermal idling is needed for the CM_CORE to idle and in-turn
PD_CORE to attain low power states.

These Bandgap-Thermal notifiers will be called only when next power state
for the CORE is less than or equal to INACTIVE and GPU power domain is OFF.

Change-Id: I42e9cc2e227f02b2237ea5770a964b114fad6c19
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>