OMAPDSS: HDMI: Dont turn ON HDMI if HPD is not High.
OMAPFB calls panel enable function for default display.
In case of HDMI as default display TV may not be present
during boot up. HPD high indicates presence of TV.
It is recommended to not turn on the HDMI if HPD is low
during bootup. This will save us power as well as
compositions to be not posted to hdmi during boot up
eventhough HDMI is not present.

Change-Id: I20a5271608fa7287ea5625f9852f0f4d552157fb
Signed-off-by: Dandawate Saket <dsaket@ti.com>