OMAP: Serial: Enable FIFO Mode and disable TX_EMPTY flag.
The original code was essentially disabling the FIFO mode by clearing the
FIFO_EN bit just before writing the FCR register. The patch allows FIFO
mode to be used.
The TX_EMPTY flag that causes interrupts to be generated when the TX Fifo is emply vs
using the TX threshold value for interrupt generation is not enabled.

This patch shows a huge improvement in audio quality when audio is streamed to
a Bluetooth headset (over UART5 for OMAP5).

Change-Id: I7096462921e7f633914038108689d6ff146f569a
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>