HACK: I2C: OMAP: Fix usecount mismatch from calls to runtime PM APIs
After coming out of suspend/resume in the resume
path we sometimes see that L4PER retention counter
does not increment. This is root caused to the I2C
driver pm_runtime_get_sync returning -EACCES in
omap_i2c_xfer function.

The issue seems that after PM runtime is disabled
in suspend and before it is enabled in resume path
we sometimes get a I2C request and in it's xfer
function get sync fails. The get_sync call increases
the usecount even though it fails and due to this
when get_sync/put_sync is called second time around,
put_sync will decrement the usecount but does not
disable the module mode as usecount > 0 resulting in
L4PER retention counter never incrementing

On failure of get sync call before returning a call
to put sync ensures that usage count is balanced
and L4PER can enter retention.

Re-visit: Once the framework takes care of decreasing
the counter if get_sync fails

Change-Id: I93b0ae90c89c2258ba1f508294b754e4db709d0c
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Madhusudan C M <c-madhu@ti.com>