ARM: OMAP: clockdomain: enable usecounting for SWSUP only domains
The patch 23d4e9a "ARM: OMAP3+: clockdomain: Enable usecounting for
clocks" enables usecounting for clockdomains which are controlled
from clock framework.
However, while doing so, it enables usecounting for clockdomains which
are HW Controllable as well. This will lead to a situation where even
if a module is marked as MODULEMODE_HWCTRL, usecounting for such
modules will be done thanks to below sequence of code.

 omap_device_enable_hwmods -> usecount=0; CLKTRCTRL=HW_AUTO
 |-_enable
   |-omap_hwmod_enable(a:MODULEMODE_HWCTRL)
     |-clkdm_hwmod_enable()	-> usecount=0; CLKTRCTRL=SW_WKUP
     |-clk_enable()		    -> usecount=1; CLKTRCTRL=SW_WKUP (main clock)
     |-clkdm_allow_idle()	-> CLKTRCTRL=HW_AUTO

As result, clkdm_usecount=1 after HWMOD enabling, which should have
remained 0.
This can be fixed by enabling usecounting for a clockdomains from
clock framework only if clockdomain is manually controlled
(CLKDM_CAN_SWSUP).

Change-Id: I808d9614d4046a124f8217b2f6514ac3ceb55b50
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>