ARM: OMAP3+: VC: vcbypass needs to wait for valid to clear
PRM_VC_VAL_BYPASS.VALID bit validates the bypass command. It is
automatically cleared by H/W either after getting the acknowledge back
from the SMPS or if an error occurred. But now in omap_vc_bypass_scale()
function current code is waiting for the invert of this.
Hence, fix the same.

Change-Id: I0434d756f2974048e66c63f51263c810220a2a6c
[grygorii.strashko@ti.com: forward port from K3.0]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>