HACK: AVS: SmartReflex/OMAP: work around pm_runtime limitations for suspend
pm_runtime APIs do not function while in platform_ops for the final stages
of suspend. This is due to the fact that pm_runtime auto-disables the devices
as part of the core handling of suspend operations.
Unfortunately, this implies that AVS operations need to be stopped prior to
runtime framework switching off it's functionality. This implies many 100s of
millisecond delay (potentially more depending on suspend handlers from
other drivers as well which may delay suspend operations) prior to system
achiving the targetted state.

This unfortunately is not acceptable from H/w perspective as the
voltage monitoring loop will be disabled beyond reasonable period of time.

However, attempting to do a pm_runtime operation as part of final stages of
suspend will also not function due to the fact that pm_runtime is disabled.

So, we HACK around this limitation by introducing suspend flag and selectively
invoking pm_runtime operations when *not in suspended* state (aka in final
stages of suspend) OR invoking omap_device APIs(and bypass run_time framework)
in the final stages of suspend.

This control is provided by platform data provided operations that can
selectively choose the API set to use.

NOTE: a slight increase of function call depth is implied by this change,
however, it guarentees to us that clk are appropriately disabled allowing us
to achieve deepest power state possible for OMAP.

This does not impact normal operational routines such as CPUIdle/DVFS as
pm_runtime framework is active during this period in time.

[nm@ti.com: HACK modified with previous upstream-posted version]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>