GPIO: OMAP: Disable GPIOs only when going to suspend mode
Disable GPIOs when going to suspend mode and keep them
in HW auto during the CPU idle path so they will transition
when L4PER is going to transition.

The runtime suspend handler of GPIO takes care of changing
level triggered to edge since GPIO block can detect only
edges in idle. Since in idle path we do not disable GPIO
runtime suspend handler will not be called. To ensure that
we do not miss level events put logic in idle handler to
take care of this scenario.

Change-Id: I2ad9254e3f57e6dfe326ce457e3cbb53027c766c
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>