tmp006: send case temperature updates to TI-TFW
On OMAP5 sEVM, there is tmp006 IR based temperature sensor
to monitor the case temperature. Previous to this patch, there
was only a debugfs entry to assist profiling and characterisation.
With this patch temperature reports are updated to TI-TFW so that
Case Governor/policy can take corrective actions to reduce the Case temperature.

Change-Id: I44a920580a4e6960b36bd7e4ec466b12cd6293a9
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>