OMAP4+: HSI HACK: work around pm_runtime limitations for suspend
pm_runtime APIs do not function during suspend.
This is due to the fact that pm_runtime auto-disables the devices
as part of the core handling of suspend operations.

So when HSI ISR is called for an IO PAD wakeup, pm_runtime can't be called.
We HACK around this limitation by calling omap_device_runtime_resume_helper
and omap_device_runtime_suspend_helper functions.

Change-Id: I65a73afb3e1972188bf8479aba2aca573543f67d
Signed-off-by: Cedric Poignet <c-poignet@ti.com>