hwspinlock: provide implementation to reset a spinlock
hwspinlocks shared between the mpu and a remote processor can be
left in a locked state in the event of a device exception on the
remote processor have taken a hwspinlock.

The mpu being the master of the remote processor, is the best
choice to try and unlock the hwspinlock.

A new API __hwspin_lock_reset is added to allow the mpu to unlock
or reset a hwspinlock.

Change-Id: Id28e03e3ce54741f522b066ba487bf78e740c041
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>