omap5: bandgap: clean-up redundant code for notifier
Since the PM notifier call is modified to invoke the
notifiers only in cpu-idle path, removing the redundant
code from the driver(which was handling, this notifier being
called in suspend path).

The commit - f72b66f ARM: OMAP4+: PM: Call thermal idle notifer only in idle path -
is doing this partially.

This patch removes the rest of code which is no longer required.

Change-Id: I148bf9922ffdc1bec2b81e163af0d84a7c74db63
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>