omap: remoteproc: forcibly reset hwspinlocks held by a rproc
Any hwspinlock held by a remote processor cannot be acquired by
the host or other remote processors if the remote processor has
crashed or been stopped with the hwspinlock held.

These hwspinlocks are forcibly reset whenever a remoteproc has
stopped or is going through a recovery. This ensures that the
hwspinlocks are brought to a sane state before any of the
processors can use them.

Change-Id: I53d73bd63743cb1d61b14bd946e62be15102dcd8
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>