omap: remoteproc: handle hwspinlock custom resource
The remote processor defines a hwspinlock custom resource in its
resource table to expose the state of hwspinlocks that it uses.
The resource provides information about the number of hwspinlocks
as well as the state variable. The state variable is an array of
integers with each bit-field storing the information of a spinlock.
The number of spinlocks dictates the number of integer elements
used for bit-fields.

This hwspinlock resource information needs to be handled and stored
for cleanup during recovery.

Change-Id: I4f740a6f5882eda92ca44dab0be1d56cacc47dcc
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>