ARM: OMAP3+: powerdomain: Check and update counters on PD transition to ON
-1-
omap_set_pwrdm_state switches powerstate of powerdomains and increments
transition counters in all cases except when powerdomain is already ON.
A powerdomain could have transitioned from a lower power state to ON
due to a wakeup interrupt or if a dependent domain becomes ON (example:
MPU, CPUx).

So even, if power domain was found ON, check prev status and increment
counters of prev state, if needed.

-2-
Use pwrdm_state_high2low_counter_update to update counters when low
power switch is turned ON. There is to standardize usage of counter
update functions.

Change-Id: Ic63ff38c869106e25063f4bad7f3660215d76ba3
Signed-off-by: Mahesh Renduchintala <mahesh@ti.com>