Revert "ARM: OMAP5: GPIO: do module reset coming from context loss"
This reverts commit b78b955ac2d7949c5dcabb5d9474b9eb658335c1.

This patch along with commit 94be942fa4879d8acd5ce64bd6d4d468c41e4e26
introduced GPIO reset as described by h/w bug COBRA-1.0BUG00134.
However on further investigations and discussions with h/w team it is
concluded that it is not required to reset the GPIO modules. This
issue was observed in first place due to IRQSTATUS being cleared when
interrupts are disabled. When this was fixed in simulation and software
by ensuring wakeup is cleared when interrupts are enabled this issue is
not seen anymore.

Change-Id: Ia6955b9106d2a56d05b6c0e0454d1793c264c378
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>