GPIO: OMAP: common runtime handlers for suspend & idle path
In GPIO driver earlier idle notifiers would call runtime
get/put sync APIs to disable the module. However in the
current implementation we keep the GPIO module in auto
mode instead of disabling in idle path but do an explicit
disable in suspend path. Due to this runtime suspend/resume
handlers are not called in idle path.
However we need the functionality in runtime suspend/resume
handlers so instead of duplicating the code create common
hanlders called from both these paths.

Also GPIO driver needs to ensure context is restored before
isolation is removed and this can happen only in idle resume
notifier.However since it is too early in idle resume notifier
and runtime APIs are not accesible we skip restoring context
here and do it when runtime resume is called. Once framework
API is available we can enable module, do access and disable
it again.

Change-Id: I0ab48c64d72e45d0ccae4c2790adeac5e300850a
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>