ARM: OMAP2+: hwmod: Introduce HWMOD_ACCESS_DISABLED flag.
OMAP5432 introduced a case when IP block is present on die but not connected
to balls. It is recommended to disable such IP block and prevent any
further access by S/W.
Introduce a HWMOD_ACCESS_DISABLED flag and omap_hwmod_register_flags() function.
The omap_hwmod_register_flags() is intended to be called early in boot before
the clock framework is initialized. Ideally, it should be called right after
omap_hwmod_register().
It will not have any effect if called not in _HWMOD_STATE_REGISTERED state.

Change-Id: I6846ac61e36dc4f9088c9348ed5dba8b98b36d87
Signed-off-by: Oleg Matcovschi <oleg.matcovschi@ti.com>