ASoC: omap-abe: Ignore suspend for legacy DMic DAI link
Inherited from commit 938fff07fd9c4cd4cde6b9a8554d0655c39718fd
in p-android-omap-3.0 to adapt to p-linux-omap-3.4 as part of
migration from k3.0 to k3.4

Commit 52ba67b (ASoC: Force all DAPM contexts into the same bias
state) has a negative effect on card's bias level due to widgetless
contexts. Bias level for widgetless contexts is calculated based
on the DAPM event. For example, for RESUME event, it bumps the
card's bias level up to power-on state. This is suboptimal and will
cost additional power transitions to the sound card.

In the meantime the ASoC framework fixes that problem, the side
effect of it can be prevent by making the corresponding DAI link
(DMic legacy) as ignore-suspend capable, so that SUSPEND/RESUME
DAPM events are not sent because of DMic legacy DAI link.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>

Change-Id: I36add7f37b9bd760a3231828334d9f261fc310b6
Signed-off-by: Ranganath Krishnan <ranganath@ti.com>