USB: EHCI: OMAP: switch back to the native ehci_irq handler
The only reason of implementing custom ehci_irq handler
was specific implementation of SW workaround for the
"OMAP5430-1.0BUG01648:No PCD IRQ generated upon remote wakeup"
HW bug. It was implemented using routed IRQ from wakeups
to the ehci hwmod ISRs and so there might be a situation,
when ehci_irq handler was called on IO wakeup event and
so the Host might be suspended.
Now when a workaround for this HW bug is reimplemented,
there is no more need in custom ehci_irq handler since
ehci generates IRQs when the clocks are enabled and
the Host is not suspendned

Change-Id: Ib058117c4c88f174fb7539d9a395bd5eb237cf0f
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>