GPIO: OMAP: HACK: always enable debounce clock for gpio WA
For the h/w bug (OMAP5430-1.0BUG01667) GPIO WA to be
effective L4PER needs to be ON before WA is applied
and also before it is removed. Without this the
glitch would still occur.

This is obvious if WA is removed in resume from idle
since L4PER can be idling as it is too early in
resume. It is not seen if WA is removed in runtime
resume at a later stage since L4PER might be ON by
that time. Since there is a common handler for idle
and suspend and we need to call it in resume from
idle for non EMIF isolation to be effective this
patch ensures that debounce clock for atleast one
GPIO bank is switched on when applying or removing
the WA.

This is labled as HACK since we are forcing GPIO8
debounce clock ON only for the GPIO WA and hence it
is under ES1.0 check since the h/w is fixed in ES2.0

Change-Id: I64fb85d2e43eed528f4270d6707e5cc51bfd189b
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Nicole Chalhoub <n-chalhoub@ti.com>