OMAPDSS: panel tc358765 initialization delay after power-on
This change adds delay for 100ms between powering on of display
panel tc358765 and its initialization. This delay is needed,
because time to time registers of the panel may be not initialized
directly after power-on. The value for delay was found
experimentally. Delay in tc358765_hw_reset() function (100ms) also
plays role.

Change-Id: I889a498787490adb004d0324a44fa44047278f6a
Signed-off-by: Mykhailo Denysiuk <x0172934@ti.com>