ARM: omap: gpio: set wakeupenable bits when entering idle modes
THe pad wakeupenable bits are required to wake through a gpio when
in OFF or OSWR.  If the wakeupenable bits are set when the controller
is not in idle, a PRCM and a GPIO interrupt will occur for every
interrupt.  Set the wakeupenable bits for every active gpio
interrupt when entering any gpio idle mode.

Change-Id: Iada59f0ac0d03b0d67c4c15a372c1b9043cb9962
Tested-by: Nicole Chalhoub <n-chalhoub@ti.com>
Signed-off-by: Colin Cross <ccross@android.com>
[huzefank@ti.com: ported from K3.0]
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
[t-kristo@ti.com: added support for reconfiguring io chain after re-mux]
Signed-off-by: Tero Kristo <t-kristo@ti.com>