ARM: OMAP3+: powerdomain: clear prev status after updating counters
In order to maintain accuracy of transition counters, clear prev
status of the powerdomain after counters are incremented. If not
done, counter updates could be inaccurate. Below are a few scenarios.

-1-
MPU counters can be incremented either by CPU0 or CPU1 during
cpuidle. The counter update is based on comparison between
prev state and current state. So if prev state is not cleared
after being processed to one CPU to update counters, then
the second CPU would update the counters the same way as
the first CPU leading to incorrect counters.

-2-
Clocks in a clockdomain could cause powerdomain transitions, when
clocks get enabled, disabled, idled etc. On the event of a powerdomain
transition, functions to update counters are called from clockdomain
framework. If prev bit is not cleared, then counters will be updated
multiple times for a single powerdomain transition, on every
other clock operation in that powerdomain.

Summary of changes
Clear prev state bit, when the powerdomain transitioned to a new
state.

Change-Id: Ifec2cc187ec4ac89a265a204d2af8d21f0893f94
Signed-off-by: Mahesh Renduchintala <mahesh@ti.com>