AVS: SmartReflex: remove redundant volt parameter from sr_enable
sr_enable takes volt parameter, which is not really used in
sr_enable. sr_enable figures the configuration using
omap_voltage_get_curr_vdata which is a better representation
of OPP voltage data that it is operating on. So remove
volt as a parameter, and fix minor debug information
for the same as well.

Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>