OMAP5: Palmas: Fix ID detection interrupt handler
This patch fixes the ID detection interrupt handler invoked when a type A micro
USB cable is connected.
The PALMAS_USB_ID_INT_LATCH_CLR register needs to be written, to clear the
source of the interrupt. If this is not done the PALMAS_USB_ID_INT_LATCH_SET
register is not updated for subsequent interrupts.

Change-Id: Iebe016d5ad74060e41b21a72af8f692ba3c06d7d
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>