ARM: OMAP4+: Do not post transitions for all powerdomains in cpuidle
Currently, during active usecases (display on), cpuidle posts
transitions for all power domains. This has two issues

-1-
posting transition for all power domains is a non-critical
bookkeeping task. This can cause delays when cpus are woken up
to attend to a have to a critical event

-2-
In-addition, there are power domains that transition to
different power states even when cpus are not idling. Posting
transitions for all power domains still does not give an
accurate account of power state transitions in the system

Remove posting transitions for all powerdomains.
Transition counters are updated when omap_set_pwrdm_state
is called for CPUx, MPU and Core

Signed-off-by: Mahesh Renduchintala <mahesh@ti.com>

Minor-fix only Merge Conflicts:

	arch/arm/mach-omap2/omap-mpuss-lowpower.c

Change-Id: I74531bb94a2654aef5bc5e49158462ccdf0921fd
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>