OMAP5: Palmas: Avoid spurious VBUS IRQ delay
When a micro A USB cable is plugged in , a.k.a the controller is operating
in host mode, the spurious vbus irqs create an undesired delay of
"PALMAS_INT3_LINE_STATE_TIME". This is unnecessary, so if the link status
indicates host mode (ID grounded), the patch allows the interrupt handler
routine to exit immediately.

Change-Id: If069bcee9cf99cf4d2154ef1c86fafb69d0f8d0f
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>