DWC3: Host: Change probe sequence
This patch changes the way the host xhci stack is initialized.
The xhci stack is loaded when then a USB peripheral is connected and
unlaoded when the peripheral is disconnected.
The patch works around an issue seen with the event mailbox where it is
observed that the mailboxing intended to change the DWC3 mode to host does
not work.
This approach is taken for the following reasons:-
- The phys need to be powered on for the host probe to work.
- Whenever use of the DWC3 IP is initiated in gadget mode the core is reset,
  demanding a reload of the xhci stack.

This patch will need to be revisted when xhci stack is re-
architected.

Change-Id: Ibfb9295623a91630729ecdf460555694ef32186f
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>