ARM OMAP2+ OPP: Perform OPP registration as part of device build
Currently an OPP is registered as part of OPP Init and
an opp is not added if the corresponding device for that OPP is not
available. It's quite possible that by the time OPP init is called,
many of the devices may not be available. This will lead to those OPPs
not registered with DVFS layer.
It is better to register an OPP for a device when the device is built.
This will avoid dependency on device creation sequence with OPP layer.
As part of this change, based OPP init is moved as part of early init so
that OPP table information is available when devices are getting built.

Change-Id: Id9a085f21f21d1371f86d18e591350fd6c520b98
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>
[ivan.khoronzhuk@ti.com: moved omap_init_opp_table to __init]
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>