ARM: dts: omap4-panda: Add sound node for HDMI
This sound node represents the HDMI audio card.
Also, it adds a label to the level shifter so that it can be
identified in sound_hdmi.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>

This node aims to represent the connection between the OMAP HDMI IP
and the TPD12S015 level-shifter chip. Thus, the hame is updated to
better reflect that.

Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>