HACK: ARM: OMAP2+: register secure device for use by security driver
Register SMC driver as platform device and
export PM functions needed by SMC driver as part of platform
data - This is known to work with OMAP5 ES1.0, but the strategy
can be made to work with multiple SoCs as well.
Tested against:
* SMC version AH01.01p2.36793.

This is still a hack considering the APIs being exported out,
but it allows us to scope the fixes that need to be
eventually done.

OMAP5 introduces an expected base and offset.
Board files need to initiate the defaults using
omap5_secure_workspace_addr_default() call in their .reserve
prior to calling omap_reserve.
This could have been handled using a cpu_is check, however
IO maps are not ready at the point in time of reserve, nor
is cpu_is check ready yet, so an explicit call is mandated
per platform supported.

If boards do not meet the assumptions of the TI default
release of PPA, the board files may use the following API
instead of omap5_secure_workspace_addr_default to set the
the platform specific address and size as needed prior to
omap_reserve is invoked as part of .reserve handler.
void omap_secure_set_secure_workspace_addr(phys_addr_t secure_workspace_addr,
                                          phys_addr_t secure_workspace_size);

Change-Id: Icdd82ff5446b5f45a52316038c72684f3da5f47e
Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>