ARM: OMAP4: powerdomain: Add supported INACTIVE power state
On OMAP4, one can explicitly program INACTIVE as the power state of
the logic area inside the power domain. Technically PD state programmed
to ON and if all the clock domains within the PD are idled, is
equivalent to PD programmed to INACTIVE and all the clock domains within
the PD are idled. There won't be any power difference in above two.

This patch enables the support of power domain INACTIVE state
programming which is supported on OMAP4.

Change-Id: Iaf08b29e262c7cdd10df8a07589dd6ef8cafeff3
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Cc: Rajendra Nayak <rnayak@ti.com>
Cc: Benoit Cousson <b-cousson@ti.com>
Acked-by: Axel Haslam <axelhaslam@ti.com>
[grygorii.strashko@ti.com - Forward port from K3.0]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>