remoteproc: added rproc_find_fw_section() fxn
The rproc_find_fw_section() searches for the given section name
and returns the addr of the section name. It looks at all
S-Sections names inside the ELF image format. It is used for
finding the .resource_table and .version sections. It can also
be used to locate the .certificate section for secure playback.

Change-Id: I5d11c254ba43669f1938f812151e39c361ea44c8
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>