ARM: OMAP5: mux: Mux changes required to support OMAP5432 ES1.0
The OMAP5432 ES1.0 is offered in a 754-ball, comparing to OMAP5430 980-ball. Also,
some of the OMAP5432 device pads are not available on the package balls.
As a consequence:
• Stacked mode and LPDDR2 memories are not supported. The OMAP5432 supports DDR3/DDR3L
memory interface through bottom ballout.
• Some of the interfaces offered in the die are not available.
• Some of the interfaces offered in the die are restricted.
• Some of the device system pins are not supported.

Disable mux of unsupported pins for OMAP5432 ES1.0.
Change omap5_mux_init() logic to depend on omap revision instead of package flags.

Change-Id: I899669639a2d3be6f929bb557b780576274a5983
Signed-off-by: Oleg Matcovschi <oleg.matcovschi@ti.com>