TI thermal framework: tmp102 sensor: get domain name from plat data
In order to proper setup several instances of tmp102 temperature sensors,
it is required to configure the thermal domain per device instance.

This patch changes the probe function to read the domain name and
the device name from the platform data. This way, the domain can
be set per board definition.

Change-Id: Ic4506ea59f99f1ea1cbc9700e85558d32e904508
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>