board-omap5evm: add device instance for tmp102 close to battery
This patch adds a I2C device which represents the tmp102 temperature
sensor close to battery in OMAP5 sEVM. This change also
defines different thermal domains for each tmp102 instance,
by setting the proper platform data field.

As the tmp102 close to the battery is not used to compute
extrapolated temperature inside OMAP, there is no need
to setup the slope and const values in its platform data.

Change-Id: Ic1aa54145910fc66b51ea952f6b758bb7d3d46de
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>