TI thermal framework: improve logic on case governor
The current code was updating the cooling level and the
thresholds if temperature would be slowly increasing.

This patch changes the logic in a way so that it makes
sure that:
a. System will be restarted when max cooling level is reached
and a proper message is present in log output.
b. The cooling devices are reset when temperature is
bellow sys_threshold_cold.
c. There is no action if we are in between subzones.
d. Debug messages are printed with correct printk level.

Change-Id: I420577fa6619fd490004a15e8873906ee3a37e21
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>