ARM: OMAP3+: Adaptive Body-Bias structures & data
Starting with OMAP36XX, some voltage domains have an LDO for biasing
voltage to the transistors within that domain.

This LDO has three modes of operation.  The first is Forward Body-Bias
mode (FBB) which boosts performance of transistors at high OPPs by
providng a positive voltage bias.  This comes at a cost of power.

The second mode is Reverse Body-Bias or RBB.  This mode provides a
negative voltage bias which saves on static leakage at lower OPPs.

Finally ABB can also be bypassed (the default state) in which case it
will just follow the voltage of the VP/VC.

This patch introduces the data structures needed to represent the ABB
LDOs in the voltage layer, and populates the appropriate data for 3630
and OMAP4.

Not all voltage domains have an ABB LDO; on OMAP36XX the CORE voltage
domain does not have an ABB ldo, and there are none on OMAP34xx.  In
such cases the voltage data will be marked with OMAP_ABB_NO_LDO.

Change-Id: If51fe172c78b2e20933c80a5f62709bd6fbcfad9
Signed-off-by: Mike Turquette <mturquette@ti.com>
Signed-off-by: Mike Turquette <mturquette@linaro.org>
[andrii.tseglytskyi@ti.com: rebased to K3.4, add OMAP5 data, RBB support]
Signed-off-by: Andrii Tseglytskyi <andrii.tseglytskyi@ti.com>