TI thermal framework: case domain: Improve performance
By enhancing the way we perform the cooling, we can
improve the performance and still keep the temperature
in controlled zone.

This patch changes the cooling strategy in cpufreq cooling
device for the case domain so that we consider
the cooling level will be going up and down
whenever temperature is greater than sys threshold hot.

Change-Id: I7f1f9617e16379e43f4c32aba5781887d5295371
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>