MFD: twl6040: PDM_UL errata
When twl6040 enters into Sleep/Deep-Sleep mode, the value of PDMUP could be
at logic 1 (VDDVIO) state. In this state, the pulldown resistor on OMAP side
results in a current drain.

Since the logic 1 reset value cannot be modified on the TWL6040, the solution
to stop the current drain is to disable the pulldown resistor on the input of
the PDM buffer of OMAP.

Since the suggested workaround is not in twl6040 side, but in McPDM pin mux
side, the actual implementation of the workaround is not in twl6040 MFD itself.

Change-Id: If375a32f01fe2dbe116cefcccafb8cc818747308
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>