OMAP4+: HSI: Add the IRQ wake enabling API call
As part of the interrupt framework this enable_irq_wake() function
has to be called if the IRQ awakes the system.

It's the case for our HSI IRQ handlers since they are used from K3.4
for a wakeup on CA_WAKE PAD thanks to the IO ring PAD wakeup
mechanism (Daisy Chain using PRCM HW IRQ).

Change-Id: Ifcf3356b6002c9250cf4ae0d45a3fd2b6ac64ea3
Signed-off-by: Cedric Poignet <c-poignet@ti.com>