HACK:OMAPDSS:Restart LCD_EN in case GO bit reset timeout triggers
Incase of UFPSR not coming back to HFUISR, switch OFF & ON the LCD_EN
bit, so that if DISPC has gone to a HALT state, will be re-configured back.

This patch is a WA for this issue & the actual fix to be root-caused on why
the DISPC has been halted even after enabling LCD_EN bit.

Change-Id: I5fa9144f731734957be985dc8858733b3de31924
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>