OMAPDSS: DSS: Add MEM2MEM manager mode
This change adds memory-to-memory (M2M) mode with using of the
manager. The main idea is to connect write back (WB) pipeline to
the manager (overlay, in terms of technical reference manual)
which is not in use (physical interface is turned-off) for using
advantages of hardware manager in M2M mode. In this case several
pipelines can be connected to the manager.

Change-Id: Ib17a041337c96c23b89e6beb69b62cfe6e7c8e97
Author: Mykhailo Denysiuk <x0172934@ti.com>
Co-author: Sergiy Kibrik <sergiikibrik@ti.com>
Signed-off-by: Mykhailo Denysiuk <x0172934@ti.com>
Signed-off-by: Artem Rudenko <artemrudenko@ti.com>