rtc: rtc-twl: set the device as wakeup capable
Since the rtc-twl device is wakeup capable it should be
set accordingly in the driver.
Without this the alarmtimer interface in kernel 3.4 fails
to add the rtc device and subsequently the android alarm
interface (alarm-dev) is not able to set time/alarm through
TWL RTC module. As a result the date/time settings are not
preserved on subsequent boots.

Change-Id: I5f39bfcae9e29ecd42e28f5f5c0969b93d1ea6d3
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>