TI thermal framework: fake case sensor: fix pdata setup
This patch changes the probe function of the fake case sensor
driver so that we always register to the case domain, instead
of using the .source_domain pdata entry. The .source_domain will
be used only for reading the sensor temperature.

Also the work queue sampling period was been fetch from average_number
instead of report_delay_ms. This patch changes the probe so that it
uses the proper pdata field

Change-Id: I7c2935c315ccf7b89c52d85a9d65e5fbb1a2c4cc
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>