remoteproc: add msleep while waiting for the firmware
The remoteproc driver repeatedly performs a request_firmware in
a thread for 120 seconds to wait for the udev to become available
and return a valid firmware data. A delay of 1 sec has been added
between each request to avoid causing an uevent socket saturation
and starvation of uevents from other drivers due to this continous
generation of uevents from the remoteproc driver.

Change-Id: I81d3b35a472edb2fa9c1976f3f48be54c8d06a96
Reported-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>